First version, too inconvenient for the client (read below).
The decoding process uses a single [decode] function that provides
input data and decodes the next element at the same time. The encoding
process uses a single [encode] function that provides the output
buffer and the element to encode at the same time. This results in
ugly and complex code for the client. Too much index fiddling and
cumbersome control flow. The result is also very unconvenient for
those that want simple blocking behavior.

The two aspects, providing input/output data, and decoding/encoding
should be kept seperate. This should also allow a better unification
of blocking/non-blocking control flow.